refac: change restaurant type as enum
