+++ internal/repository/meal_repository.go
@@ -8,8 +8,7 @@ import (
 	"strings"
 	"time"
 
-	"github.com/School-meal-lover/backend/app/internal/models"
-
+	"github.com/School-meal-lover/backend/internal/models"
 	"github.com/google/uuid"
 )
 
@@ -21,36 +20,39 @@ func NewMealRepository(db *sql.DB) *MealRepository {
 	return &MealRepository{db: db}
 }
 
-// 레스토랑 조회
-func (r *MealRepository) GetRestaurantByName(name string) (*models.Restaurant, error) {
-	query := `SELECT id FROM restaurants WHERE name = $1`
-
-	var restaurant models.Restaurant
-	err := r.db.QueryRow(query, name).Scan(&restaurant.ID)
-	if err != nil {
-		return nil, fmt.Errorf("restaurant not found for name %s: %w", name, err)
-	}
-
-	restaurant.Name = name
-	return &restaurant, nil
-}
-
 // 주차 정보 삽입
-func (r *MealRepository) InsertWeek(startDate time.Time, restaurantID string) (string, error) {
+func (r *MealRepository) InsertWeek(startDate time.Time, restaurant models.RestaurantType) (string, error) {
 	weekID := uuid.New().String()
 	query := `
-        INSERT INTO weeks (id, start_date, restaurants_id, created_at, updated_at)
+        INSERT INTO weeks (id, start_date, restaurant, created_at, updated_at)
         VALUES ($1, $2, $3, now(), now())
-        RETURNING id`
+				RETURNING id`
 
 	var insertedID string
-	err := r.db.QueryRow(query, weekID, startDate, restaurantID).Scan(&insertedID)
+	err := r.db.QueryRow(query, weekID, startDate, restaurant).Scan(&insertedID)
 	if err != nil {
 		return "", fmt.Errorf("failed to insert week: %w", err)
 	}
 
 	return insertedID, nil
 }
+func (r *MealRepository) FindOrCreateWeek(startDate time.Time, restaurant models.RestaurantType) (string, error) {
+	var weekID string
+	findQuery := `SELECT id FROM weeks WHERE start_date = $1 AND restaurant = $2`
+	
+	err := r.db.QueryRow(findQuery, startDate, restaurant).Scan(&weekID)
+	
+	if err == nil {
+		log.Printf("Found existing week ID: %s for start date %s", weekID, startDate.Format("2006-01-02"))
+		return weekID, nil
+	}
+
+	if err == sql.ErrNoRows {
+		return r.InsertWeek(startDate, restaurant)
+	}
+	return "", fmt.Errorf("error while trying to find week: %w", err)
+}
+
 
 // 식사 정보 삽입
 func (r *MealRepository) InsertMeal(meal *models.Meal) (string, error) {
@@ -73,6 +75,22 @@ func (r *MealRepository) InsertMeal(meal *models.Meal) (string, error) {
 	}
 
 	return insertedID, nil
+}
+func (r *MealRepository) FindOrCreateMeal(meal *models.Meal) (string, error) {
+    var mealID string
+    findQuery := `SELECT id FROM meals WHERE weeks_id = $1 AND date = $2 AND meal_type = $3`
+
+    err := r.db.QueryRow(findQuery, meal.WeekID, meal.Date, meal.MealType).Scan(&mealID)
+
+    if err == nil {
+        return mealID, nil
+    }
+
+    if err == sql.ErrNoRows {
+        return r.InsertMeal(meal)
+    }
+
+    return "", fmt.Errorf("error finding or creating meal: %w", err)
 }
 
 func (r *MealRepository) HandleRepositoryError(err error, notFoundCode, notFoundMessage string) (*models.RestaurantMealsResponse, error) {
@@ -97,10 +115,10 @@ func (r *MealRepository) InsertMenuItems(menuItems []models.MenuItem) error {
 		return fmt.Errorf("failed to begin transaction: %w", err)
 	}
 	defer func() {
-		if err := tx.Rollback(); err != nil  && err != sql.ErrTxDone {
+		if err := tx.Rollback(); err != nil && err != sql.ErrTxDone {
 			log.Printf("failed to rollback transaction: %v", err)
-		}	
-		}()
+		}
+	}()
 
 	stmt, err := tx.Prepare(`
 				INSERT INTO menu_items (id, meals_id, category, name, name_en, price, created_at, updated_at)
@@ -129,29 +147,19 @@ func (r *MealRepository) InsertMenuItems(menuItems []models.MenuItem) error {
 	return tx.Commit()
 }
 
-// 레스토랑 정보 조회
-func (r *MealRepository) GetRestaurantInfo(restaurantID string) (*models.RestaurantInfo, error) {
-	restaurant := &models.RestaurantInfo{}
-	query := `SELECT id, name, COALESCE(name_en, '') FROM restaurants WHERE id = $1`
-
-	err := r.db.QueryRow(query, restaurantID).Scan(
-		&restaurant.ID, &restaurant.Name, &restaurant.NameEn)
-
-	return restaurant, err
-}
-func (r *MealRepository) GetWeekInfo(restaurantId, date string) (*models.WeekInfo, error) {
+func (r *MealRepository) GetWeekInfo(restaurant models.RestaurantType, date string) (*models.WeekInfo, error) {
 	week := &models.WeekInfo{}
 	var startDate time.Time
 	query := `
 		SELECT id, start_date
         FROM weeks 
-        WHERE restaurants_id = $1 
+        WHERE restaurants = $1 
         AND $2 >= start_date 
         AND $2 <= start_date + INTERVAL '6 days'
         ORDER BY start_date DESC
         LIMIT 1`
 
-	err := r.db.QueryRow(query, restaurantId, date).Scan(&week.ID, &startDate)
+	err := r.db.QueryRow(query, restaurant, date).Scan(&week.ID, &startDate)
 	if err != nil {
 		return nil, fmt.Errorf("failed to get week by date: %w", err)
 	}
@@ -193,7 +201,7 @@ func (r *MealRepository) GetMealsData(weekID string) ([]*models.DayMeals, *model
 		var date time.Time
 		var price float64
 
-		err := rows.Scan(&mealID, &date, &dayOfWeek, &mealType,&category,&menuID, &menuName, &menuNameEn, &price)
+		err := rows.Scan(&mealID, &date, &dayOfWeek, &mealType, &category, &menuID, &menuName, &menuNameEn, &price)
 		if err != nil {
 			return nil, nil, err
 		}
@@ -220,11 +228,11 @@ func (r *MealRepository) GetMealsData(weekID string) ([]*models.DayMeals, *model
 		// 메뉴 아이템 넣기
 		if menuID != "" {
 			menuItem := &models.MenuItemResponse{
-				ID:     menuID,
+				ID:       menuID,
 				Category: category,
-				Name:   menuName,
-				NameEn: menuNameEn,
-				Price:  price,
+				Name:     menuName,
+				NameEn:   menuNameEn,
+				Price:    price,
 			}
 			mealMap[mealKey].MenuItems = append(mealMap[mealKey].MenuItems, menuItem)
 			totalMenuItems++
@@ -233,16 +241,16 @@ func (r *MealRepository) GetMealsData(weekID string) ([]*models.DayMeals, *model
 	var orderedDays []*models.DayMeals
 	for _, dayMeal := range mealsByDay {
 		orderedDays = append(orderedDays, dayMeal)
-  }
-	sort.Slice(orderedDays, func(i,j int) bool {
-    dateI, errI := time.Parse(("2006-01-02"), orderedDays[i].Date)
-    dateJ, errJ := time.Parse(("2006-01-02"), orderedDays[j].Date)
-    if errI != nil || errJ != nil {
-      log.Printf("failed to parse date: %v, %v", errI, errJ)
-      return orderedDays[i].Date < orderedDays[j].Date
-    }
-    return dateI.Before(dateJ)
-  })
+	}
+	sort.Slice(orderedDays, func(i, j int) bool {
+		dateI, errI := time.Parse(("2006-01-02"), orderedDays[i].Date)
+		dateJ, errJ := time.Parse(("2006-01-02"), orderedDays[j].Date)
+		if errI != nil || errJ != nil {
+			log.Printf("failed to parse date: %v, %v", errI, errJ)
+			return orderedDays[i].Date < orderedDays[j].Date
+		}
+		return dateI.Before(dateJ)
+	})
 
 	summary := &models.MealsSummary{
 		TotalDays:      len(orderedDays),
@@ -254,47 +262,47 @@ func (r *MealRepository) GetMealsData(weekID string) ([]*models.DayMeals, *model
 }
 
 func (r *MealRepository) GetMealIDByWeekDateAndType(weekID, date, mealType string) (string, error) {
-    var mealID string
-    query := `
+	var mealID string
+	query := `
         SELECT id FROM meals 
         WHERE weeks_id = $1 AND date = $2 AND meal_type = $3
         LIMIT 1
     `
-    err := r.db.QueryRow(query, weekID, date, mealType).Scan(&mealID)
-    return mealID, err
+	err := r.db.QueryRow(query, weekID, date, mealType).Scan(&mealID)
+	return mealID, err
 }
 
 func (r *MealRepository) GetMenuItemsByMealIDOrdered(mealID string) ([]models.MenuItem, error) {
-    query := `
+	query := `
         SELECT id, meals_id, category, name, name_en, price
         FROM menu_items
         WHERE meals_id = $1
         ORDER BY id ASC
     `
-    rows, err := r.db.Query(query, mealID)
-    if err != nil {
-        return nil, err
-    }
-    defer rows.Close()
+	rows, err := r.db.Query(query, mealID)
+	if err != nil {
+		return nil, err
+	}
+	defer rows.Close()
 
-    var items []models.MenuItem
-    for rows.Next() {
-        var item models.MenuItem
-        if err := rows.Scan(&item.ID, &item.MealID, &item.Category, &item.Name, &item.NameEn, &item.Price); err != nil {
-            return nil, err
-        }
-        items = append(items, item)
-    }
-    return items, nil
+	var items []models.MenuItem
+	for rows.Next() {
+		var item models.MenuItem
+		if err := rows.Scan(&item.ID, &item.MealID, &item.Category, &item.Name, &item.NameEn, &item.Price); err != nil {
+			return nil, err
+		}
+		items = append(items, item)
+	}
+	return items, nil
 }
 
 func (r *MealRepository) UpdateMenuItemNameEn(menuItemID, nameEn string) error {
-    _, err := r.db.Exec(`
+	_, err := r.db.Exec(`
         UPDATE menu_items
         SET name_en = $1
         WHERE id = $2
     `, nameEn, menuItemID)
-    return err
+	return err
 }
 func (r *MealRepository) UpdateMenuItemsEnglishNameBatch(items []models.MenuItem) error {
 	if len(items) == 0 {
@@ -316,4 +324,3 @@ func (r *MealRepository) UpdateMenuItemsEnglishNameBatch(items []models.MenuItem
 	_, err := r.db.Exec(query, args...)
 	return err
 }
-
